remove isAdapters and move types/journals to internal
Fixes #315
