+++ packages/core/src/new-api/internal/deployer.ts
@@ -18,7 +18,6 @@ import {
 import { Artifact, ArtifactResolver } from "../types/artifact";
 import { DeployConfig, DeploymentResult } from "../types/deployer";
 import { DeploymentLoader } from "../types/deployment-loader";
-import { Journal } from "../types/journal";
 
 import { Batcher } from "./batcher";
 import { ExecutionEngine } from "./execution/execution-engine";
@@ -36,6 +35,7 @@ import {
   DeploymentExecutionState,
   ExecutionStateMap,
 } from "./types/execution-state";
+import { Journal } from "./types/journal";
 import { TransactionLookupTimer } from "./types/transaction-timer";
 import { assertIgnitionInvariant } from "./utils/assertions";
 import { getFuturesFromModule } from "./utils/get-futures-from-module";
