+++ packages/core/src/new-api/internal/deployment-loader/file-deployment-loader.ts
@@ -3,8 +3,8 @@ import path from "path";
 
 import { Artifact, BuildInfo } from "../../types/artifact";
 import { DeploymentLoader } from "../../types/deployment-loader";
-import { Journal } from "../../types/journal";
 import { FileJournal } from "../journal/file-journal";
+import { Journal } from "../types/journal";
 
 export class FileDeploymentLoader implements DeploymentLoader {
   public journal: Journal;
