+++ packages/core/src/new-api/internal/journal/file-journal.ts
@@ -2,7 +2,7 @@
 import fs, { closeSync, constants, openSync, writeFileSync } from "fs";
 import { parse } from "ndjson";
 
-import { Journal, JournalableMessage } from "../../types/journal";
+import { Journal, JournalableMessage } from "../types/journal";
 
 import { deserializeReplacer } from "./utils/deserialize-replacer";
 import { logJournalableMessage } from "./utils/log";
