+++ packages/core/src/new-api/internal/journal/memory-journal.ts
@@ -1,4 +1,4 @@
-import { Journal, JournalableMessage } from "../../types/journal";
+import { Journal, JournalableMessage } from "../types/journal";
 
 import { deserializeReplacer } from "./utils/deserialize-replacer";
 import { logJournalableMessage } from "./utils/log";
