+++ packages/core/test/new-api/helpers.ts
@@ -12,12 +12,15 @@ import {
   OnchainStatuses,
 } from "../../src/new-api/internal/types/execution-state";
 import {
+  Journal,
+  JournalableMessage,
+} from "../../src/new-api/internal/types/journal";
+import {
   DeploymentResult,
   DeploymentResultContract,
   DeploymentResultContracts,
 } from "../../src/new-api/types/deployer";
 import { DeploymentLoader } from "../../src/new-api/types/deployment-loader";
-import { Journal, JournalableMessage } from "../../src/new-api/types/journal";
 
 export const exampleAccounts: string[] = [
   "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
